refactor(cli): give the audit-line formatter its own module
